Move edit logic to parent.
